Jarvis/requirements.txt
PyQt6
PyQt6-WebEngine
requests
pvporcupine
faster-whisper
edge-tts
sounddevice
numpy
python-dotenv
colorama
pyaudio
